+++ effective-gear/common/src/main/java/com/chimericdream/effectivegear/mixin/EG$EnderpearlItemMixin.java
@@ -1,7 +1,6 @@
 package com.chimericdream.effectivegear.mixin;
 
 import com.chimericdream.effectivegear.ModInfo;
-import net.minecraft.core.component.DataComponents;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.Identifier;
 import net.minecraft.resources.ResourceKey;
@@ -18,8 +17,8 @@ public class EG$EnderpearlItemMixin {
     @Unique
     private static ResourceKey<TrimMaterial> ENDER_PEARL_TRIM_ID = ResourceKey.create(Registries.TRIM_MATERIAL, Identifier.fromNamespaceAndPath(ModInfo.MOD_ID, "ender_pearl"));
 
-    @ModifyVariable(method = "<init>(Lnet/minecraft/world/item/Item$Properties;)V", at = @At("HEAD"), ordinal = 1)
+    @ModifyVariable(method = "<init>(Lnet/minecraft/world/item/Item$Properties;)V", at = @At("HEAD"), ordinal = 0)
     private static Item.Properties eg$addTrimMaterialComponent(Item.Properties properties) {
-        return properties.delayedHolderComponent(DataComponents.PROVIDES_TRIM_MATERIAL, ENDER_PEARL_TRIM_ID);
+        return properties.trimMaterial(ENDER_PEARL_TRIM_ID);
     }
 }