+++ effective-gear/CHANGELOG.md
@@ -1,1 +1,5 @@
 ### Unreleased changes
+
+#### New Features
+
+* Added Ender pearls as an armor trim material